Add files via upload
Abstract:
In agriculture, finding plant diseases is a crucial 
activity that is important for ensuring crop health 
and increasing productivity. In this project, we 
provide an in-depth analysis of the use of a 
Convolutional Neural Network (CNN) for plant 
disease detection. This project utilizes a dataset that 
includes a sizable number of photos showing both 
healthy plants and plants with various diseases. The 
dataset is enhanced and pre-processed to improve 
the ability to be generalized the model. 
Convolutional layers, pooling layers, and fully 
connected layers are used in the creation and 
training of a CNN architecture to extract useful 
characteristics from the images.
The model's parameters are optimized during the 
training phase using stochastic gradient descent and 
a properly selected learning rate. Over-fitting is 
avoided by using regularization techniques like 
dropout and weight decay. The model's efficacy in 
identifying plant diseases can be evaluated using a 
variety of factors, including accuracy, precision, 
recall, and F1-score. To compare the proposed CNN 
model to current methods for detecting plant 
diseases, a comparative analysis is done. The 
model's advantages and disadvantages are 
examined, with an emphasis on the model's 
potential for precise disease detection and its 
applicability for practical applications.
The project's results provide a practical and 
dependable method that advances the study of 
plant disease detection. The created CNN model 
exhibits encouraging outcomes and is highly 
accurate in recognizing plant illnesses. The results 
open the door for the use of automated disease 
detection systems in agriculture, enabling early 
diagnosis and quick response to lessen the effects of 
plant diseases and improve crop management 
techniques.